requirements.txt
# Introduction to Artificial Intelligence (2026)
# Student: Denzel Nyarko | Index: 10022300153
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
pymupdf>=1.23.0
groq
python-dotenv>=1.0.0
requests>=2.28.0
